+++ play/src/front/Chat/Components/Room/CreateRoomOrFolderOption.svelte
@@ -1,0 +1,61 @@
+<script lang="ts">
+    import { openModal } from "svelte-modals";
+    import LL from "../../../../i18n/i18n-svelte";
+    import CreateRoomModal from "./CreateRoomModal.svelte";
+    import CreateFolderModal from "./CreateFolderModal.svelte";
+    import RoomOption from "./RoomMenu/RoomOption.svelte";
+    import { IconLogout, IconSquarePlus } from "@wa-icons";
+
+    export let parentID: string | undefined = undefined;
+    let optionButtonRef: HTMLButtonElement | undefined = undefined;
+    let optionRef: HTMLDivElement | undefined = undefined;
+    let hideFolderOptions = true;
+
+    function toggleSpaceOption() {
+        if (optionButtonRef === undefined) {
+            return;
+        }
+        if (optionRef === undefined) {
+            return;
+        }
+        const { bottom, right } = optionButtonRef.getBoundingClientRect();
+        optionRef.style.top = `${bottom}px`;
+        optionRef.style.left = `${right}px`;
+        hideFolderOptions = !hideFolderOptions;
+    }
+    function closeMenuAndOpenCreateRoom() {
+        hideFolderOptions = true;
+        openModal(CreateRoomModal, {
+            parentID,
+        });
+    }
+    function closeMenuAndOpenCreateSpace() {
+        hideFolderOptions = true;
+        openModal(CreateFolderModal, {
+            parentID,
+        });
+    }
+</script>
+
+<button
+    data-testid="openCreateRoomModalButton"
+    class="tw-p-0 tw-m-0 tw-text-gray-400"
+    bind:this={optionButtonRef}
+    on:click|preventDefault|stopPropagation={toggleSpaceOption}
+>
+    <IconSquarePlus font-size={16} />
+</button>
+<div
+    on:mouseleave={toggleSpaceOption}
+    bind:this={optionRef}
+    class="tw-absolute tw-bg-black/90 tw-rounded-md tw-p-1 tw-z-[1] tw-w-max"
+    class:tw-absolue={optionButtonRef !== undefined}
+    class:tw-hidden={hideFolderOptions}
+>
+    <RoomOption IconComponent={IconLogout} title={$LL.chat.createRoom.title()} on:click={closeMenuAndOpenCreateRoom} />
+    <RoomOption
+        IconComponent={IconLogout}
+        title={$LL.chat.createFolder.title()}
+        on:click={closeMenuAndOpenCreateSpace}
+    />
+</div>
